finalizando o cenário de cadastro

+++ backend/src/controllers/HelpersController.ts
@@ -1,4 +1,5 @@
 import { Request, Response } from "express"
+import FoodTruck from "../models/FoodTruck"
 
 import User from '../models/User'
 
@@ -8,6 +9,12 @@ export default {
 
         const {instagram} = req.query
 
+        const user = await User.findOne({instagram: instagram})
+
+        if(!user)
+            return res.status(204).end()
+
+        await FoodTruck.deleteMany({suggested_by: user._id})   
         await User.deleteMany({instagram: instagram})
 
         return res.status(204).end()
